+++ web_01_2024-main/Interdisciplinar/src/service/TutorService.ts
@@ -17,8 +17,14 @@ export class TutorService{
 
     async atualizarTutor(tutorData: any): Promise<Tutor> {
         const { id, name, telefone, email, cidade, bairro, rua, numero } = tutorData;
+        const idNumber = parseInt(id, 10);
 
-        const tutor = Tutor.getInstance(id, name, telefone, email, cidade, bairro, rua, numero)
+        const tutor = Tutor.getInstance(idNumber, name, telefone, email, cidade, bairro, rua, numero)
+
+        const TutorExiste = await this.tutorRepository.filterTutor(tutor.id)
+        if(TutorExiste.length == 0){
+            throw new Error("Tutor não encontrado");
+        }
 
         await this.tutorRepository.updateTutor(tutor);
         console.log("Service - Update ", tutor);
@@ -27,15 +33,21 @@ export class TutorService{
 
     async deletarTutor(tutorData: any): Promise<Tutor> {
         const { id, name, telefone, email, cidade, bairro, rua, numero } = tutorData;
+        const idNumber = parseInt(id, 10);
 
-        const tutor = Tutor.getInstance(id, name, telefone, email, cidade, bairro, rua, numero)
+        const tutor = Tutor.getInstance(idNumber, name, telefone, email, cidade, bairro, rua, numero)
+
+        const TutorExiste = await this.tutorRepository.filterTutor(tutor.id)
+        if(TutorExiste.length == 0){
+            throw new Error("Tutor não encontrado");
+        }
 
         await this.tutorRepository.deleteTutor(tutor);
         console.log("Service - Delete ", tutor);
         return tutor; 
     }
 
-    async filtrarTutor(tutorData: any): Promise<Tutor> {
+    async filtrarTutor(tutorData: any): Promise<Tutor[]> {
         const idNumber = parseInt(tutorData, 10);
 
         const tutor =  await this.tutorRepository.filterTutor(idNumber);
